+++ examples/functions/algolia-document-sync/index.ts
@@ -1,16 +1,17 @@
+import {env} from 'node:process'
+
 import {type DocumentEvent, documentEventHandler} from '@sanity/functions'
 import {algoliasearch} from 'algoliasearch'
-import process from 'node:process'
 
-const {ALGOLIA_APP_ID, ALGOLIA_WRITE_KEY} = process.env
+const {ALGOLIA_APP_ID = '', ALGOLIA_WRITE_KEY = ''} = env
 
 export const handler = documentEventHandler(async ({event}: {event: DocumentEvent}) => {
   const {_id, title, hideFromSearch} = event.data
 
-  const algolia = algoliasearch(ALGOLIA_APP_ID || '', ALGOLIA_WRITE_KEY || '')
+  const algolia = algoliasearch(ALGOLIA_APP_ID, ALGOLIA_WRITE_KEY)
 
   try {
-    // We are assuming you already have an algolia instance steup with an index called 'posts'
+    // We are assuming you already have an algolia instance setup with an index called 'posts'
     // addOrUpdateObject documentation: https://www.algolia.com/doc/libraries/javascript/v5/methods/search/add-or-update-object/?client=javascript
     await algolia.addOrUpdateObject({
       indexName: 'posts',
@@ -24,5 +25,4 @@ export const handler = documentEventHandler(async ({event}: {event: DocumentEven
     console.error('Error syncing to Algolia:', error)
     throw error
   }
-  return
 })