+++ packages/sanity/src/core/config/types.ts
@@ -415,12 +415,6 @@ export interface PluginOptions {
     components?: StudioComponentsPluginOptions
   }
 
-  /**
-   * Config for the Sanity Media Library asset source integration.
-   * @beta
-   */
-  mediaLibrary?: MediaLibraryConfig
-
   /** @beta @hidden */
   i18n?: LocalePluginOptions
   search?: {
@@ -470,6 +464,12 @@ export interface PluginOptions {
   announcements?: {
     enabled: boolean
   }
+
+  /**
+   * Config for the Sanity Media Library asset source integration.
+   * @beta
+   */
+  mediaLibrary?: DefaultPluginsWorkspaceOptions['mediaLibrary']
 }
 
 /** @internal */
@@ -531,6 +531,10 @@ export interface WorkspaceOptions extends SourceOptions {
    * @internal
    */
   releases?: DefaultPluginsWorkspaceOptions['releases']
+  /**
+   * @internal
+   */
+  mediaLibrary?: DefaultPluginsWorkspaceOptions['mediaLibrary']
   apps?: AppsOptions
 
   /**
@@ -919,7 +923,7 @@ export interface Source {
    * Config for the Sanity Media Library asset source integration.
    * @beta
    */
-  mediaLibrary?: MediaLibraryConfig
+  mediaLibrary?: WorkspaceOptions['mediaLibrary']
 }
 
 /** @internal */
@@ -1066,6 +1070,7 @@ export type DefaultPluginsWorkspaceOptions = {
      */
     limit?: number
   }
+  mediaLibrary?: MediaLibraryConfig
 }
 
 /**